+++ workspaces/jenkins/plugins/jenkins/src/components/Router.tsx
@@ -34,18 +34,21 @@ export const isJenkinsAvailable = (entity: Entity) =>
   Boolean(entity.metadata.annotations?.[JENKINS_ANNOTATION]) ||
   Boolean(entity.metadata.annotations?.[LEGACY_JENKINS_ANNOTATION]);
 
-export const Router = (props: { columns?: TableColumn<Project>[] }) => {
+export const Router = (props: {
+  title?: string;
+  columns?: TableColumn<Project>[];
+}) => {
   const { entity } = useEntity();
 
   if (!isJenkinsAvailable(entity)) {
     return <MissingAnnotationEmptyState annotation={JENKINS_ANNOTATION} />;
   }
 
-  const columns = props.columns;
+  const { title, columns } = props;
 
   return (
     <Routes>
-      <Route path="/" element={<CITable columns={columns} />} />
+      <Route path="/" element={<CITable title={title} columns={columns} />} />
       <Route path={`/${buildRouteRef.path}`} element={<DetailedViewPage />} />
       <Route path={`/${jobRunsRouteRef.path}`} element={<JobRunsTable />} />
     </Routes>